Add ‘Phonetic Transcription’ feature
Insertion of phonetic transcription to Anki notes was not provided by
the add-on, so vocal pronunciation was the sole means to learn
pronunciation of a word. This ability, however, if is provided, can
ease language learning.

Add the feature by extracting phonetic transcription of each
requested word from XML response, labeling different pronunciations
by their part of speech, and inserting them in the user defined
field. Develop in a way that it can function not only within the main
insertion procedure of AutoDefine but also autonomously as are served
for asking for either definition or vocal pronunciation. In other
words, consider also a distinct button and shortcut so that users who
only concern about written pronunciation of the words on their Anki
notes can also take advantage of this add-on.