Remove inappropriate newline at the end of definition
Not only doesn't Python newline help on Anki cards, but it also
produces trouble. If a user inserts definition of a word on a field,
by deleting the definition and inserting it again via AutoDefine,
a disturbing blank line will appear. This is definitely bothersome as
it appears unexpectedly. Remove Python newline on strings that appear
on Anki cards and just stick to html newline as it perfectly does the
charm.