Remove newline from vocal pronunciation
If a user wants to insert just vocal pronunciation, it is better not
to add unwanted newline to the end. This way they will not face with
appearance difference on their Anki card if they only add the sound.
We can consider other approaches to improve the arrangement for
multiple kinds of data insert.